pyproject.toml
[project]
name = "youtube-podcast-analyser"
version = "0.1.0"
description = "Summarizes Youtube Video"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-multipart>=0.0.9",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "audioop-lts>=0.2.2",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "defusedxml==0.7.1",
    "distro==1.9.0",
    "exceptiongroup==1.3.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "jiter==0.10.0",
    "openai==1.102.0",
    "openpyxl>=3.1.5",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydub>=0.25.1",
    "python-dotenv==1.1.1",
    "requests==2.32.5",
    "sniffio==1.3.1",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "youtube-transcript-api==1.2.2",
    "yt-dlp==2025.8.27",
    "groq>=0.31.0",
]
